backend/requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.4.0
celery==5.3.6
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
whitenoise==6.7.0
drf-spectacular==0.27.2
python-decouple==3.8
waitress==3.0.1
